requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
alembic==1.9.4
    # via flask-migrate
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via flask
colorama==0.4.6
    # via click
flask==2.2.3
    # via
    #   -r requirements.in
    #   flask-migrate
    #   flask-sqlalchemy
flask-migrate==4.0.4
    # via -r requirements.in
flask-sqlalchemy==3.0.3
    # via
    #   -r requirements.in
    #   flask-migrate
greenlet==2.0.2
    # via sqlalchemy
idna==3.4
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
mako==1.2.4
    # via alembic
markupsafe==2.1.2
    # via
    #   jinja2
    #   mako
    #   werkzeug
mistune==2.0.5
    # via -r requirements.in
mysqlclient==2.1.1
    # via -r requirements.in
python-dotenv==0.21.1
    # via -r requirements.in
requests==2.28.2
    # via -r requirements.in
sqlalchemy==2.0.4
    # via
    #   alembic
    #   flask-sqlalchemy
typing-extensions==4.5.0
    # via sqlalchemy
urllib3==1.26.14
    # via requests
waitress==2.1.2
    # via -r requirements.in
werkzeug==2.2.3
    # via flask
